functionalities completed

+++ vegge/views.py
@@ -11,38 +11,26 @@ from time import sleep
 
 # Create your views here.
 def recipes(request):
-    if request.method == 'POST':
-        # Only process add recipe if the add button was clicked
-        if 'name' in request.POST:
-            data = request.POST
-            print("POST data:", data)
-            name = data.get('name')
-            description = data.get('description')
-            image = request.FILES.get('image')
-            if name:
-                Recipe.objects.create(
-                    name=name,
-                    description=description,
-                    image=image
-                )
-            else:
-                print("Name is missing!")
-            return redirect('recipes')
-        # Otherwise, ignore POST (it's from Delete/Edit)
-        return redirect('recipes')
     queryset = Recipe.objects.all()
     if request.GET.get('search'):
        queryset = queryset.filter(name__icontains=request.GET.get('search'))
     context = {'recipes': queryset}
     return render(request, 'recipes.html', context)
+    
 
 def delete_recipe(request, id):
+    if not request.user.is_authenticated:
+        messages.error(request, "You must be logged in to delete a recipe!")
+        return redirect('login_page')
     recipe = Recipe.objects.get(id=id)
     recipe.delete()
     return redirect('recipes')
 
 def edit_recipe(request, id):
     recipe = Recipe.objects.get(id=id)
+    if not request.user.is_authenticated:
+        messages.error(request, "You must be logged in to edit a recipe!")
+        return redirect('login_page')
     if request.method == 'POST':
         data = request.POST
         name = data.get('name')
@@ -110,3 +98,27 @@ def logout_page(request):
     else:
         messages.error(request, "You are not logged in!")
         return redirect('login_page')
+    
+def add_recipe(request):
+    if not request.user.is_authenticated:
+        messages.error(request, "You must be logged in to add a recipe!")
+        return redirect('login_page')
+    if request.method == 'POST':
+        if 'name' in request.POST:
+            data = request.POST
+            print("POST data:", data)
+            name = data.get('name')
+            description = data.get('description')
+            image = request.FILES.get('image')
+            if name:
+                Recipe.objects.create(
+                    name=name,
+                    description=description,
+                    image=image
+                )
+            else:
+                print("Name is missing!")
+            return redirect('recipes')
+        return redirect('recipes')
+    # For GET requests, render the form
+    return render(request, 'add_recipe.html')
